Go: Move identify environment code to separate file
